# apps/api/api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "HTTP API for the raffle stress-test laboratory.",
        "title": "Raffle Stress Test API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/v1",
    "paths": {
        "/admin/raffles": {
            "post": {
                "security": [
                    {
                        "AdminBearer": []
                    }
                ],
                "description": "Admin only. Does not insert ticket rows until the raffle is opened.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Create a draft raffle",
                "parameters": [
                    {
                        "description": "Raffle",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateRaffleRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.RaffleResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/raffles/{id}": {
            "patch": {
                "security": [
                    {
                        "AdminBearer": []
                    }
                ],
                "description": "Opening a raffle inserts ticket numbers 1..total_tickets once. Closing keeps tickets and orders.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Change raffle status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Raffle ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Status",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateRaffleStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.RaffleResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/payments/webhooks": {
            "post": {
                "security": [
                    {
                        "WebhookSecret": []
                    }
                ],
                "description": "Applies a paid webhook once. Replays of the same event_id return 200 without selling again.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payments"
                ],
                "summary": "Confirm a simulated payment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Webhook secret",
                        "name": "X-Webhook-Secret",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Payment event",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.WebhookRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Empty"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/purchases": {
            "get": {
                "description": "Returns only orders for the normalized email. An unknown email is an empty list.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "purchases"
                ],
                "summary": "List purchases by email",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Customer email",
                        "name": "email",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.PurchaseListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/raffles": {
            "get": {
                "description": "Storefront catalog. status=open lists raffles accepting orders. status=closed lists concluded raffles. Drafts are omitted.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "raffles"
                ],
                "summary": "List public raffles by status",
                "parameters": [
                    {
                        "enum": [
                            "open",
                            "closed"
                        ],
                        "type": "string",
                        "description": "Public raffle status",
                        "name": "status",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.RaffleListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/raffles/{slug}": {
            "get": {
                "description": "Public storefront view. available_tickets is total minus reserved minus sold.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "raffles"
                ],
                "summary": "Get a raffle by slug",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Raffle slug",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.RaffleResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/raffles/{slug}/orders": {
            "post": {
                "description": "Reserves ticket capacity and returns a pending simulated payment. Numbers are not chosen here.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Open a checkout order",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Raffle slug",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Checkout",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateOrderRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.OrderResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.CreateOrderRequest": {
            "type": "object",
            "required": [
                "email",
                "quantity"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                }
            }
        },
        "dto.CreateRaffleRequest": {
            "type": "object",
            "required": [
                "slug",
                "ticket_price_cents",
                "title",
                "total_tickets"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "ticket_price_cents": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "total_tickets": {
                    "type": "integer"
                }
            }
        },
        "dto.Empty": {
            "type": "object"
        },
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "dto.OrderResponse": {
            "type": "object",
            "properties": {
                "amount_cents": {
                    "type": "integer"
                },
                "email": {
                    "type": "string"
                },
                "expires_at": {
                    "type": "string"
                },
                "order_id": {
                    "type": "string"
                },
                "payment": {
                    "$ref": "#/definitions/dto.PaymentResponse"
                },
                "quantity": {
                    "type": "integer"
                },
                "raffle_slug": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "tickets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.TicketNumber"
                    }
                }
            }
        },
        "dto.PaymentResponse": {
            "type": "object",
            "properties": {
                "payment_id": {
                    "type": "string"
                },
                "qr_code_payload": {
                    "type": "string"
                }
            }
        },
        "dto.PurchaseListResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "purchases": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.PurchaseResponse"
                    }
                }
            }
        },
        "dto.PurchaseResponse": {
            "type": "object",
            "properties": {
                "amount_cents": {
                    "type": "integer"
                },
                "order_id": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                },
                "raffle_slug": {
                    "type": "string"
                },
                "raffle_title": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "tickets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.TicketNumber"
                    }
                }
            }
        },
        "dto.RaffleListResponse": {
            "type": "object",
            "properties": {
                "raffles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.RaffleResponse"
                    }
                }
            }
        },
        "dto.RaffleResponse": {
            "type": "object",
            "properties": {
                "available_tickets": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "ticket_price_cents": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "total_tickets": {
                    "type": "integer"
                }
            }
        },
        "dto.TicketNumber": {
            "type": "object",
            "properties": {
                "number": {
                    "type": "string"
                }
            }
        },
        "dto.UpdateRaffleStatusRequest": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "status": {
                    "type": "string"
                }
            }
        },
        "dto.WebhookRequest": {
            "type": "object",
            "required": [
                "event_id",
                "payment_id",
                "status"
            ],
            "properties": {
                "event_id": {
                    "type": "string"
                },
                "payment_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "AdminBearer": {
            "description": "Admin token formatted as \"Bearer \u003ctoken\u003e\".",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "WebhookSecret": {
            "description": "Shared payment webhook secret.",
            "type": "apiKey",
            "name": "X-Webhook-Secret",
            "in": "header"
        }
    }
}